源代码/homework/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.4
Pillow==9.3.0
pip==21.3.1
PyMySQL==1.0.2
setuptools==60.2.0
sqlparse==0.4.3
tzdata==2022.7
wheel==0.37.1
